requirements.txt
urllib3==1.24.2
newspaper3k==0.2.8
nltk==3.4.1
python-dateutil==2.8.0
requests==2.21.0
lxml==4.3.3
numpy==1.16.3
pandas==0.24.2
beautifulsoup4==4.7.1